@@ -70,7 +70,7 @@ func (o *AddandremovefrombookmarkprompttemplateResponse) GetContentType() string
 
 func (o *AddandremovefrombookmarkprompttemplateResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
