@@ -85,7 +85,7 @@ func (o *AddandremovefromfavDocumentResponse) GetContentType() string {
 
 func (o *AddandremovefromfavDocumentResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
