@@ -86,7 +86,7 @@ func (o *AllTemplatesResponse) GetContentType() string {
 
 func (o *AllTemplatesResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
