@@ -117,7 +117,7 @@ func (o *CreateProductResponse) GetCreateProduct() *CreateProductCreateProduct {
 
 func (o *CreateProductResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
