@@ -58,7 +58,7 @@ func (o *DeleteallkindofdocumentsResponse) GetDeleteallkindofdocuments() *Delete
 
 func (o *DeleteallkindofdocumentsResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
