@@ -77,7 +77,7 @@ func (o *DeleteCustomtemplateResponse) GetDeleteCustomtemplate() *DeleteCustomte
 
 func (o *DeleteCustomtemplateResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
