@@ -77,7 +77,7 @@ func (o *DeleteWorkspaceResponse) GetDeleteWorkspacelive() *DeleteWorkspaceDelet
 
 func (o *DeleteWorkspaceResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
