@@ -117,7 +117,7 @@ func (o *GenerateCodeResponse) GetGenerateCode() *GenerateCodeGenerateCode {
 
 func (o *GenerateCodeResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
