@@ -78,7 +78,7 @@ func (o *JoinworkbookrequestrResponse) GetContentType() string {
 
 func (o *JoinworkbookrequestrResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
