@@ -70,7 +70,7 @@ func (o *ParmanentDeletePrompttemplateResponse) GetContentType() string {
 
 func (o *ParmanentDeletePrompttemplateResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
