@@ -94,7 +94,7 @@ func (o *ProcessTemplateResponse) GetContentType() string {
 
 func (o *ProcessTemplateResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
