@@ -87,7 +87,7 @@ func (o *PromptTemplatesResponse) GetContentType() string {
 
 func (o *PromptTemplatesResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
