@@ -78,7 +78,7 @@ func (o *RestoreDocuemntofalltypeContentVoiceoverImagesTranscriptCodeResponse) G
 
 func (o *RestoreDocuemntofalltypeContentVoiceoverImagesTranscriptCodeResponse) GetHeaders() map[string][]string {
 	if o == nil {
-		return nil
+		return map[string][]string{}
 	}
 	return o.Headers
 }
